requirements.txt
flask
python-dotenv
openai
pandas
python-docx
PyPDF2
openpyxl
